+++ pkg/formula/input/prompt/prompt.go
@@ -55,6 +55,7 @@ type InputManager struct {
 	prompt.InputBool
 	prompt.InputPassword
 	prompt.InputMultiselect
+	prompt.InputPath
 }
 
 func NewInputManager(
@@ -66,6 +67,7 @@ func NewInputManager(
 	inBool prompt.InputBool,
 	inPass prompt.InputPassword,
 	inMultiselect prompt.InputMultiselect,
+	inPath prompt.InputPath,
 ) formula.InputRunner {
 	return InputManager{
 		cred:               cred,
@@ -76,6 +78,7 @@ func NewInputManager(
 		InputBool:          inBool,
 		InputPassword:      inPass,
 		InputMultiselect:   inMultiselect,
+		InputPath:          inPath,
 	}
 }
 
@@ -146,7 +149,12 @@ func (in InputManager) inputTypeToPrompt(items []string, i formula.Input) (strin
 			return "", err
 		}
 		return in.List(i.Label, dl, i.Tutorial)
-	case input.Multiselect:
+	case input.PathType:
+		if items != nil {
+			return in.loadInputValList(items, i)
+		}
+		return in.InputPath.Read(i.Label)
+	case input.MultiselectType:
 		if len(items) == 0 {
 			return "", fmt.Errorf(EmptyItems, i.Name)
 		}
@@ -154,7 +162,7 @@ func (in InputManager) inputTypeToPrompt(items []string, i formula.Input) (strin
 		if err != nil {
 			return "", err
 		}
-		return strings.Join(sl, "|"), nil
+		return strings.Join(sl, input.MultiselectSeparator), nil
 	default:
 		return in.cred.Resolve(i.Type)
 	}
@@ -207,18 +215,21 @@ func (in InputManager) persistCache(formulaPath, inputVal string, input formula.
 	}
 }
 
-func (in InputManager) loadInputValList(items []string, input formula.Input) (string, error) {
+func (in InputManager) loadInputValList(items []string, i formula.Input) (string, error) {
 	newLabel := DefaultCacheNewLabel
-	if input.Cache.Active {
-		if input.Cache.NewLabel != "" {
-			newLabel = input.Cache.NewLabel
+	if i.Cache.Active {
+		if i.Cache.NewLabel != "" {
+			newLabel = i.Cache.NewLabel
 		}
 		items = append(items, newLabel)
 	}
 
-	inputVal, err := in.List(input.Label, items, input.Tutorial)
+	inputVal, err := in.List(i.Label, items, i.Tutorial)
 	if inputVal == newLabel {
-		return in.textValidator(input)
+		if i.Type == input.PathType {
+			return in.InputPath.Read(i.Label)
+		}
+		return in.textValidator(i)
 	}
 
 	return inputVal, err