+++ pkg/formula/repo.go
@@ -18,21 +18,35 @@ package formula
 
 import (
 	"sort"
+	"time"
 
 	"github.com/ZupIT/ritchie-cli/pkg/git"
 )
 
-const RepoCommonsName = RepoName("commons")
+const (
+	RepoCacheTime   = 24 * time.Hour
+	RepoCommonsName = RepoName("commons")
+)
 
 type Repo struct {
-	Provider    RepoProvider `json:"provider"`
-	Name        RepoName     `json:"name"`
-	Version     RepoVersion  `json:"version"`
-	Url         string       `json:"url"`
-	Token       string       `json:"token,omitempty"`
-	Priority    int          `json:"priority"`
-	IsLocal     bool         `json:"isLocal"`
-	TreeVersion string       `json:"tree_version"`
+	Provider      RepoProvider `json:"provider"`
+	Name          RepoName     `json:"name"`
+	Version       RepoVersion  `json:"version"`
+	Url           string       `json:"url"`
+	Token         string       `json:"token,omitempty"`
+	Priority      int          `json:"priority"`
+	IsLocal       bool         `json:"isLocal"`
+	TreeVersion   string       `json:"tree_version"`
+	LatestVersion RepoVersion  `json:"latest_version"`
+	Cache         time.Time    `json:"cache"`
+}
+
+func (r Repo) CacheExpired() bool {
+	return r.Cache.Before(time.Now())
+}
+
+func (r *Repo) UpdateCache() {
+	r.Cache = time.Now().Add(RepoCacheTime)
 }
 
 func (r Repo) EmptyVersion() bool {
@@ -142,10 +156,12 @@ type RepositoryListWriter interface {
 	RepositoryWriter
 }
 
-type RepositoryListWriteCreator interface {
-	RepositoryLister
-	RepositoryWriter
+type RepositoryCreateWriteListDetailDeleter interface {
 	RepositoryCreator
+	RepositoryWriter
+	RepositoryLister
+	RepositoryDetail
+	RepositoryDeleter
 }
 
 type RepositoryListUpdater interface {